% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_coordinates_xanthos_reference}
\alias{data_coordinates_xanthos_reference}
\title{data_coordinates_xanthos_reference}
\format{
R tibble
}
\source{
Xanthos reference coordinates for 67420 cells. Original in xanthos/example/input/reference/coordinates.csv
}
\usage{
data_coordinates_xanthos_reference
}
\description{
data_coordinates_xanthos_reference
}
\examples{
\dontrun{
 library(im3components);
 im3components::coordinates_xanthos_reference
}
}
\keyword{datasets}
