% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_NCAR_gcam}
\alias{data_NCAR_gcam}
\title{data_NCAR_gcam}
\format{
R tibble
}
\source{
GCAM data system: ./input/gcamdata/inst/extdata/gcam-usa/NCAR_SSO2_pop_state.csv
}
\usage{
data_NCAR_gcam
}
\description{
Exisitng NCAR popultation data by US state from GCAM for SSP2.
}
\details{
Available at: ./input/gcamdata/inst/extdata/gcam-usa/NCAR_SSO2_pop_state.csv
}
\examples{
\dontrun{
 library(im3components);
 im3components::data_NCAR_gcam
}
}
\keyword{datasets}
