% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_NCAR_raw}
\alias{data_NCAR_raw}
\title{data_NCAR_raw}
\format{
R tibble
}
\source{
Jiang et al. 2020, Raw Data: http://doi.org/10.5281/zenodo.3956703
}
\usage{
data_NCAR_raw
}
\description{
Raw data from Jiang et al. 2020 for population by US state. Sources in details below.
}
\details{
\itemize{
\item Raw data from: Jiang, Leiwen, et al. 2020 "Population scenarios for US states consistent with shared
socioeconomic pathways." Environmental Research Letters 15.9 (2020): 094097;
\item Paper: https://iopscience.iop.org/article/10.1088/1748-9326/aba5b1/pdf
\item Raw data downloaded from: http://doi.org/10.5281/zenodo.3956703 and unzipped.
\item Relevant file: statepop-v0.1.0/IMMM-SFA-statepop-61c8fff/inputs/AllStatesProjection.csv
}
}
\examples{
\dontrun{
 library(im3components);
 im3components::data_NCAR_raw
}
}
\keyword{datasets}
