% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ncdf_resampled_wrf_xanthos}
\alias{data_ncdf_resampled_wrf_xanthos}
\title{data_ncdf_resampled_wrf_xanthos}
\format{
R tibble
}
\source{
IM3 WRF data set and Xanthos reference data
}
\usage{
data_ncdf_resampled_wrf_xanthos
}
\description{
data_ncdf_resampled_wrf_xanthos
}
\examples{
\dontrun{
 library(im3components);
 im3components::ncdf_resampled_wrf_xanthos
}
}
\keyword{datasets}
