% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_gcam.R
\name{gcam_modelx_sum}
\alias{gcam_modelx_sum}
\title{Sample function for the IM3 components R package}
\usage{
gcam_modelx_sum(x, y)
}
\arguments{
\item{x}{numeric Some numeric value}

\item{y}{numeric Some numeric value}
}
\value{
numeric The sum of x and y
}
\description{
Sums two numeric inputs and returns the value
}
\examples{
gcam_modelx_sum(1.4, 2.2)

}
\author{
Chris R. Vernon
}
