% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_to_gcam.R
\name{pop_gcam_process}
\alias{pop_gcam_process}
\title{pop_gcam_process}
\usage{
pop_gcam_process(
  data_NCAR_raw = im3components::data_NCAR_raw,
  data_NCAR_gcam = im3components::data_NCAR_gcam
)
}
\arguments{
\item{data_NCAR_raw}{Default = im3components::data_NCAR_raw. Must be an R table.
Can replace with updated data set when available. See format of data head(im3components::data_NCAR_raw).}

\item{data_NCAR_gcam}{Default = im3components::data_NCAR_gcam. Must be an R table.
Can replace with updated data set when available. See format of data head(im3components::data_NCAR_gcam).}
}
\value{
NCAR_SSP_pop_state.csv
}
\description{
Script to process raw NCAR population data in format similar to existing input file: gcam-usa/NCAR_SSP2_pop_state
Raw data from: Jiang, Leiwen, et al. 2020 "Population scenarios for US states consistent with shared
socioeconomic pathways." Environmental Research Letters 15.9 (2020): 094097;
data: http://doi.org/10.5281/zenodo.3956703 ; Paper: https://iopscience.iop.org/article/10.1088/1748-9326/aba5b1/pdf
Raw data downloaded from: http://doi.org/10.5281/zenodo.3956703 and unzipped.
Relevant file: statepop-v0.1.0/IMMM-SFA-statepop-61c8fff/inputs/AllStatesProjection.csv
Also uses NCAR data from GCAM to get correct format. .input/gcamdata/inst/extdata/gcam-usa/NCAR_SSO2_pop_state.csv
Output file: NCAR_SSP_pop_state.csv
}
