% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_xanthos_process.R
\name{resample_wrf_hourly_to_month}
\alias{resample_wrf_hourly_to_month}
\title{resample_wrf_hourly_to_month}
\source{
Method based on https://rpubs.com/markpayne/132500
}
\usage{
resample_wrf_hourly_to_month(
  ncdf_path = NULL,
  target_grid = NULL,
  params = NULL,
  aggregation_method = c("mean"),
  out_dir = "output",
  save = FALSE,
  ncdf_resampled = NULL
)
}
\arguments{
\item{ncdf_path}{Default = NULL. Path to netcdf file or folder containing '.nc' files.}

\item{target_grid}{Default = NULL. Path to target_grid file .csv or a dataframe}

\item{params}{Default = NULL. Params from the netcdf file to resample.
For Xanthos params= c("RAINC","Q2","PSFC","T2","GLW","SWDOWN","V10","U10")}

\item{aggregation_method}{Default = c("mean"). Aggregation method to aggregate param values over hours.
Can be a vector of the same length as params with either "sum" or "mean" corresponding to each param. If single value given then
the same method will be applied to all parameters.
For Xanthos params aggregation_method = c("sum","mean","mean","mean","mean","mean","mean","mean")}

\item{out_dir}{Default = "output". Output directory to save outputs if save = T.}

\item{save}{Default = FALSE. Whether to save files.}

\item{ncdf_resampled}{(Optional) Default = NULL. A resampled data set for WRF to Xanthos is available in the package as: im3components::data_ncdf_resampled_wrf_xanthos}
}
\value{
Dataframe of resampled data
}
\description{
Resample a netcdf4 file to new lat/lon coordinates based on
a given dataframe. Returns a long dataframe with lat lon and parameter from the netcdf chosen.
}
