% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_xanthos_resample.R
\name{wrf_xanthos_resample}
\alias{wrf_xanthos_resample}
\title{wrf_xanthos_resample}
\source{
Method based on https://rpubs.com/markpayne/132500
}
\usage{
wrf_xanthos_resample(
  ncdf_path = NULL,
  target_grid = NULL,
  params = NULL,
  aggregation_method = c("mean"),
  out_dir = "output",
  save = FALSE,
  ncdf_resampled = NULL
)
}
\arguments{
\item{ncdf_path}{Default = NULL. Path to netcdf file or folder containing '.nc' files.}

\item{target_grid}{Default = NULL. Path to target_grid file .csv or a dataframe}

\item{params}{Default = NULL. Params from the netcdf file to resample.
For Xanthos params= c("RAINC","Q2","PSFC","T2","GLW","SWDOWN","V10","U10")}

\item{aggregation_method}{Default = c("sum"). Aggregation method to aggregate param values over hours.}

\item{out_dir}{Default = "output". Output directory to save outputs if save = T.}

\item{save}{Default = FALSE. Whether to save files.}

\item{ncdf_resampled}{Default = NULL. A pre-resampled dataset for WRF to a target grid.
A pre-resampled data set for WRF to the Xanthos grid is available in the package as: im3components::data_ncdf_resampled_wrf_xanthos}
}
\value{
Dataframe of resampled data
}
\description{
This function is a wrapper for resample_wrf_hourly_to_month() to meet the naming conventions of the
im3 project. It simply calls the generic resample_wrf_hourly_to_month() function.
Resample a netcdf4 file to new lat/lon coordinates based on
a given dataframe. Returns a long dataframe with lat lon and parameter from the netcdf chosen.
}
