% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_xanthos_process.R
\name{xanthos_npy_expand}
\alias{xanthos_npy_expand}
\title{xanthos_npy_expand}
\usage{
xanthos_npy_expand(
  base_npy = NULL,
  base_year_month_start = NULL,
  target_npy = NULL,
  target_year_month_start = NULL,
  out_dir = NULL
)
}
\arguments{
\item{base_npy}{Default = NULL. Base .npy file to expand}

\item{base_year_month_start}{Default = NULL}

\item{target_npy}{Default = NULL. Target .npy file to expand to.}

\item{target_year_month_start}{Default = NULL}

\item{out_dir}{Default = NULL. Path to folder to save outputs in.}
}
\description{
This function expands a base npy file to match the dimensions of a given npy file.
Used to expand example xanthos climate data to run with wrf data for the US.
}
